Replace hardcoded resolver cache mutation with CSV cache.Source.
