+++ pkg/controller/registry/resolver/step_resolver.go
@@ -9,8 +9,6 @@ import (
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/labels"
-	"k8s.io/client-go/kubernetes"
 
 	"github.com/operator-framework/api/pkg/operators/v1alpha1"
 	"github.com/operator-framework/operator-lifecycle-manager/pkg/api/client/clientset/versioned"
@@ -37,7 +35,6 @@ type OperatorStepResolver struct {
 	subLister              v1alpha1listers.SubscriptionLister
 	csvLister              v1alpha1listers.ClusterServiceVersionLister
 	client                 versioned.Interface
-	kubeclient             kubernetes.Interface
 	globalCatalogNamespace string
 	satResolver            *SatResolver
 	log                    logrus.FieldLogger
@@ -57,15 +54,23 @@ func (pp catsrcPriorityProvider) Priority(key cache.SourceKey) int {
 	return catsrc.Spec.Priority
 }
 
-func NewOperatorStepResolver(lister operatorlister.OperatorLister, client versioned.Interface, kubeclient kubernetes.Interface,
-	globalCatalogNamespace string, provider RegistryClientProvider, log logrus.FieldLogger) *OperatorStepResolver {
+func NewOperatorStepResolver(lister operatorlister.OperatorLister, client versioned.Interface, globalCatalogNamespace string, provider RegistryClientProvider, log logrus.FieldLogger) *OperatorStepResolver {
+	cacheSourceProvider := &mergedSourceProvider{
+		sps: []cache.SourceProvider{
+			SourceProviderFromRegistryClientProvider(provider, log),
+			&csvSourceProvider{
+				csvLister: lister.OperatorsV1alpha1().ClusterServiceVersionLister(),
+				subLister: lister.OperatorsV1alpha1().SubscriptionLister(),
+				logger:    log,
+			},
+		},
+	}
 	stepResolver := &OperatorStepResolver{
 		subLister:              lister.OperatorsV1alpha1().SubscriptionLister(),
 		csvLister:              lister.OperatorsV1alpha1().ClusterServiceVersionLister(),
 		client:                 client,
-		kubeclient:             kubeclient,
 		globalCatalogNamespace: globalCatalogNamespace,
-		satResolver:            NewDefaultSatResolver(SourceProviderFromRegistryClientProvider(provider, log), catsrcPriorityProvider{lister: lister.OperatorsV1alpha1().CatalogSourceLister()}, log),
+		satResolver:            NewDefaultSatResolver(cacheSourceProvider, catsrcPriorityProvider{lister: lister.OperatorsV1alpha1().CatalogSourceLister()}, log),
 		log:                    log,
 	}
 
@@ -84,21 +89,6 @@ func (r *OperatorStepResolver) Expire(key cache.SourceKey) {
 }
 
 func (r *OperatorStepResolver) ResolveSteps(namespace string) ([]*v1alpha1.Step, []v1alpha1.BundleLookup, []*v1alpha1.Subscription, error) {
-	// create a generation - a representation of the current set of installed operators and their provided/required apis
-	allCSVs, err := r.csvLister.ClusterServiceVersions(namespace).List(labels.Everything())
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	// TODO: build this index ahead of time
-	// omit copied csvs from generation - they indicate that apis are provided to the namespace, not by the namespace
-	var csvs []*v1alpha1.ClusterServiceVersion
-	for i := range allCSVs {
-		if !allCSVs[i].IsCopied() {
-			csvs = append(csvs, allCSVs[i])
-		}
-	}
-
 	subs, err := r.listSubscriptions(namespace)
 	if err != nil {
 		return nil, nil, nil, err
@@ -106,7 +96,7 @@ func (r *OperatorStepResolver) ResolveSteps(namespace string) ([]*v1alpha1.Step,
 
 	var operators cache.OperatorSet
 	namespaces := []string{namespace, r.globalCatalogNamespace}
-	operators, err = r.satResolver.SolveOperators(namespaces, csvs, subs)
+	operators, err = r.satResolver.SolveOperators(namespaces, subs)
 	if err != nil {
 		return nil, nil, nil, err
 	}
@@ -249,4 +239,22 @@ func (r *OperatorStepResolver) listSubscriptions(namespace string) ([]*v1alpha1.
 	}
 
 	return subs, nil
+}
+
+type mergedSourceProvider struct {
+	sps    []cache.SourceProvider
+	logger logrus.StdLogger
+}
+
+func (msp *mergedSourceProvider) Sources(namespaces ...string) map[cache.SourceKey]cache.Source {
+	result := make(map[cache.SourceKey]cache.Source)
+	for _, sp := range msp.sps {
+		for key, source := range sp.Sources(namespaces...) {
+			if _, ok := result[key]; ok {
+				msp.logger.Printf("warning: duplicate sourcekey: %q\n", key)
+			}
+			result[key] = source
+		}
+	}
+	return result
 }
